// src/apispec/src/models/ActivateFunctionRevisionRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ActivateFunctionRevisionRequest
 */
export interface ActivateFunctionRevisionRequest {
    /**
     * 
     * @type {string}
     * @memberof ActivateFunctionRevisionRequest
     */
    revisionId: string;
}

/**
 * Check if a given object implements the ActivateFunctionRevisionRequest interface.
 */
export function instanceOfActivateFunctionRevisionRequest(value: object): value is ActivateFunctionRevisionRequest {
    if (!('revisionId' in value) || value['revisionId'] === undefined) return false;
    return true;
}

export function ActivateFunctionRevisionRequestFromJSON(json: any): ActivateFunctionRevisionRequest {
    return ActivateFunctionRevisionRequestFromJSONTyped(json, false);
}

export function ActivateFunctionRevisionRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ActivateFunctionRevisionRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'revisionId': json['revision_id'],
    };
}

export function ActivateFunctionRevisionRequestToJSON(json: any): ActivateFunctionRevisionRequest {
    return ActivateFunctionRevisionRequestToJSONTyped(json, false);
}

export function ActivateFunctionRevisionRequestToJSONTyped(value?: ActivateFunctionRevisionRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'revision_id': value['revisionId'],
    };
}