// src/apispec/src/models/Function.ts
export interface Function {
    /**
     * 
     * @type {string}
     * @memberof Function
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Function
     */
    teamId: string;
    /**
     * 
     * @type {string}
     * @memberof Function
     */
    createdBy?: string;
    /**
     * 
     * @type {string}
     * @memberof Function
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Function
     */
    slug: string;
    /**
     * 
     * @type {string}
     * @memberof Function
     */
    domainLabel: string;
    /**
     * 
     * @type {string}
     * @memberof Function
     */
    url?: string;
    /**
     * 
     * @type {string}
     * @memberof Function
     */
    activeRevisionId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Function
     */
    enabled: boolean;
    /**
     * 
     * @type {FunctionScalePolicy}
     * @memberof Function
     */
    scale: FunctionScalePolicy;
    /**
     * 
     * @type {Date}
     * @memberof Function
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof Function
     */
    updatedAt: Date;
}