// src/apispec/src/models/FunctionDeployRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { FunctionRevisionSpec } from './FunctionRevisionSpec';
import {
    FunctionRevisionSpecFromJSON,
    FunctionRevisionSpecFromJSONTyped,
    FunctionRevisionSpecToJSON,
    FunctionRevisionSpecToJSONTyped,
} from './FunctionRevisionSpec';
import type { FunctionScalePolicy } from './FunctionScalePolicy';
import {
    FunctionScalePolicyFromJSON,
    FunctionScalePolicyFromJSONTyped,
    FunctionScalePolicyToJSON,
    FunctionScalePolicyToJSONTyped,
} from './FunctionScalePolicy';
import type { FunctionSource } from './FunctionSource';
import {
    FunctionSourceFromJSON,
    FunctionSourceFromJSONTyped,
    FunctionSourceToJSON,
    FunctionSourceToJSONTyped,
} from './FunctionSource';

/**
 * 
 * @export
 * @interface FunctionDeployRequest
 */
export interface FunctionDeployRequest {
    /**
     * 
     * @type {string}
     * @memberof FunctionDeployRequest
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof FunctionDeployRequest
     */
    slug?: string;
    /**
     * 
     * @type {FunctionScalePolicy}
     * @memberof FunctionDeployRequest
     */
    scale?: FunctionScalePolicy;
    /**
     * 
     * @type {FunctionSource}
     * @memberof FunctionDeployRequest
     */
    source?: FunctionSource;
    /**
     * 
     * @type {FunctionRevisionSpec}
     * @memberof FunctionDeployRequest
     */
    spec?: FunctionRevisionSpec;
    /**
     * 
     * @type {boolean}
     * @memberof FunctionDeployRequest
     */
    activate?: boolean;
}

/**
 * Check if a given object implements the FunctionDeployRequest interface.
 */
export function instanceOfFunctionDeployRequest(value: object): value is FunctionDeployRequest {
    return true;
}

export function FunctionDeployRequestFromJSON(json: any): FunctionDeployRequest {
    return FunctionDeployRequestFromJSONTyped(json, false);
}

export function FunctionDeployRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): FunctionDeployRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'slug': json['slug'] == null ? undefined : json['slug'],
        'scale': json['scale'] == null ? undefined : FunctionScalePolicyFromJSON(json['scale']),
        'source': json['source'] == null ? undefined : FunctionSourceFromJSON(json['source']),
        'spec': json['spec'] == null ? undefined : FunctionRevisionSpecFromJSON(json['spec']),
        'activate': json['activate'] == null ? undefined : json['activate'],
    };
}

export function FunctionDeployRequestToJSON(json: any): FunctionDeployRequest {
    return FunctionDeployRequestToJSONTyped(json, false);
}

export function FunctionDeployRequestToJSONTyped(value?: FunctionDeployRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'slug': value['slug'],
        'scale': FunctionScalePolicyToJSON(value['scale']),
        'source': FunctionSourceToJSON(value['source']),
        'spec': FunctionRevisionSpecToJSON(value['spec']),
        'activate': value['activate'],
    };
}