// src/apispec/src/models/FunctionRevisionMount.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FunctionRevisionMount
 */
export interface FunctionRevisionMount {
    /**
     * 
     * @type {string}
     * @memberof FunctionRevisionMount
     */
    snapshotId: string;
    /**
     * 
     * @type {string}
     * @memberof FunctionRevisionMount
     */
    mountPath: string;
    /**
     * Function snapshot mounts are immutable and are materialized as read-only volumes at runtime.
     * @type {boolean}
     * @memberof FunctionRevisionMount
     */
    readOnly?: boolean;
}

/**
 * Check if a given object implements the FunctionRevisionMount interface.
 */
export function instanceOfFunctionRevisionMount(value: object): value is FunctionRevisionMount {
    if (!('snapshotId' in value) || value['snapshotId'] === undefined) return false;
    if (!('mountPath' in value) || value['mountPath'] === undefined) return false;
    return true;
}

export function FunctionRevisionMountFromJSON(json: any): FunctionRevisionMount {
    return FunctionRevisionMountFromJSONTyped(json, false);
}

export function FunctionRevisionMountFromJSONTyped(json: any, ignoreDiscriminator: boolean): FunctionRevisionMount {
    if (json == null) {
        return json;
    }
    return {
        
        'snapshotId': json['snapshot_id'],
        'mountPath': json['mount_path'],
        'readOnly': json['read_only'] == null ? undefined : json['read_only'],
    };
}

export function FunctionRevisionMountToJSON(json: any): FunctionRevisionMount {
    return FunctionRevisionMountToJSONTyped(json, false);
}

export function FunctionRevisionMountToJSONTyped(value?: FunctionRevisionMount | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'snapshot_id': value['snapshotId'],
        'mount_path': value['mountPath'],
        'read_only': value['readOnly'],
    };
}