// src/apispec/src/models/FunctionRevisionSpec.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sandbox0 API
 * Public HTTP APIs exposed by Sandbox0 regional gateways and the global gateway.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SandboxAppService } from './SandboxAppService';
import {
    SandboxAppServiceFromJSON,
    SandboxAppServiceFromJSONTyped,
    SandboxAppServiceToJSON,
    SandboxAppServiceToJSONTyped,
} from './SandboxAppService';
import type { FunctionRevisionMount } from './FunctionRevisionMount';
import {
    FunctionRevisionMountFromJSON,
    FunctionRevisionMountFromJSONTyped,
    FunctionRevisionMountToJSON,
    FunctionRevisionMountToJSONTyped,
} from './FunctionRevisionMount';

/**
 * Canonical runtime contract compiled from every function deploy mode.
 * @export
 * @interface FunctionRevisionSpec
 */
export interface FunctionRevisionSpec {
    /**
     * 
     * @type {string}
     * @memberof FunctionRevisionSpec
     */
    template: string;
    /**
     * 
     * @type {SandboxAppService}
     * @memberof FunctionRevisionSpec
     */
    service: SandboxAppService;
    /**
     * 
     * @type {Array<FunctionRevisionMount>}
     * @memberof FunctionRevisionSpec
     */
    mounts?: Array<FunctionRevisionMount>;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof FunctionRevisionSpec
     */
    envVars?: { [key: string]: string; };
}

/**
 * Check if a given object implements the FunctionRevisionSpec interface.
 */
export function instanceOfFunctionRevisionSpec(value: object): value is FunctionRevisionSpec {
    if (!('template' in value) || value['template'] === undefined) return false;
    if (!('service' in value) || value['service'] === undefined) return false;
    return true;
}

export function FunctionRevisionSpecFromJSON(json: any): FunctionRevisionSpec {
    return FunctionRevisionSpecFromJSONTyped(json, false);
}

export function FunctionRevisionSpecFromJSONTyped(json: any, ignoreDiscriminator: boolean): FunctionRevisionSpec {
    if (json == null) {
        return json;
    }
    return {
        
        'template': json['template'],
        'service': SandboxAppServiceFromJSON(json['service']),
        'mounts': json['mounts'] == null ? undefined : ((json['mounts'] as Array<any>).map(FunctionRevisionMountFromJSON)),
        'envVars': json['env_vars'] == null ? undefined : json['env_vars'],
    };
}

export function FunctionRevisionSpecToJSON(json: any): FunctionRevisionSpec {
    return FunctionRevisionSpecToJSONTyped(json, false);
}

export function FunctionRevisionSpecToJSONTyped(value?: FunctionRevisionSpec | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'template': value['template'],
        'service': SandboxAppServiceToJSON(value['service']),
        'mounts': value['mounts'] == null ? undefined : ((value['mounts'] as Array<any>).map(FunctionRevisionMountToJSON)),
        'env_vars': value['envVars'],
    };
}